backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# MongoDB async driver
motor==3.3.2
pymongo==4.6.1

# Authentication - using bcrypt directly for reliability
bcrypt==4.1.2
PyJWT==2.8.0

# Validation
pydantic[email]==2.5.3
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0
